% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyPackages.R
\name{addControlsToStudyPackage}
\alias{addControlsToStudyPackage}
\title{Add controls to a SkeletonComparativeEffectStudy package}
\usage{
addControlsToStudyPackage(
  controlConcepts,
  targetId,
  comparatorId = NULL,
  fileName = "inst/settings/NegativeControls.csv",
  type = "outcome"
)
}
\arguments{
\item{controlConcepts}{data.frame with at least 1 row (recommended > 10) containing conceptId and conceptName}

\item{targetId}{target cohort Id controls are for}

\item{comparatorId}{comparator cohort Id controls are for (optional in self controlled studies)}

\item{fileName}{path of existing negative control file or where file will be written}

\item{type}{"outcome" or "exposure" - only outcomes are currently supported by the study package}
}
\description{
See https:://github.com/ohdsi/SkeletonComparativeEffectStudy for details on creating a study
This function adds a set of negative controls to the study package to be automatically instantiated as cohorts.
Id's are determined by the study package, normally this is the same as the concept id
}
