% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvidenceExplorer.R
\name{ceExplorerModule}
\alias{ceExplorerModule}
\title{CE Explorer module}
\usage{
ceExplorerModule(
  id,
  backend,
  ingredientConceptInput = reactive({     data.frame() }),
  conditionConceptInput = reactive({     data.frame() }),
  siblingLookupLevelsInput = reactive({     0 })
)
}
\arguments{
\item{id}{string - unique namespace for module}

\item{backend}{CemConnector backend (database or Api URL)}

\item{ingredientConceptInput}{shiny::reactive that returns data.frame with headers conceptId, includeDescendants, isExcluded}

\item{conditionConceptInput}{shiny::reactive that returns data.frame with headers conceptId, includeDescendants, isExcluded}

\item{siblingLookupLevelsInput}{shiny::reactive that returns positive integer for sibling levels to lookup for condition concept mappings to CEM}
}
\description{
Shiny Module for integration of evidence for conceptsets in to shiny applications
}
