% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEMConnector.R
\docType{package}
\name{CEMConnector-package}
\alias{CEMConnector-package}
\title{CEMConnector: Common Evidence Model Connector}
\description{
RESTFul API and Database Backend connections for the OHDSI Common Evidence Model
    Built around an R Plumber api.
}
\keyword{internal}
