% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backend.R
\name{CemDatabaseBackend}
\alias{CemDatabaseBackend}
\title{CEM Database Backend Class}
\value{
data.frame of sources

data.frame of condition concept_id and concept_name

data.frame of condition concept_id and concept_name
}
\description{
An interface to the common evidence model that uses works directly with a database schema
}
\section{Super class}{
\code{\link[CemConnector:AbstractCemBackend]{CemConnector::AbstractCemBackend}} -> \code{CemDatabaseBackend}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{connection}}{ConnectionHandlder instance}

\item{\code{cemSchema}}{schema of CEM database}

\item{\code{vocabularySchema}}{OMOP vocabulary schema (must include concept and concept ancestor tables)}

\item{\code{sourceSchema}}{schema containing source_info table}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CemDatabaseBackend$new()}}
\item \href{#method-finalize}{\code{CemDatabaseBackend$finalize()}}
\item \href{#method-getConditionEvidenceSummary}{\code{CemDatabaseBackend$getConditionEvidenceSummary()}}
\item \href{#method-getConditionEvidence}{\code{CemDatabaseBackend$getConditionEvidence()}}
\item \href{#method-getIngredientEvidenceSummary}{\code{CemDatabaseBackend$getIngredientEvidenceSummary()}}
\item \href{#method-getIngredientEvidence}{\code{CemDatabaseBackend$getIngredientEvidence()}}
\item \href{#method-getRelationships}{\code{CemDatabaseBackend$getRelationships()}}
\item \href{#method-getCemSourceInfo}{\code{CemDatabaseBackend$getCemSourceInfo()}}
\item \href{#method-getSuggestedControlCondtions}{\code{CemDatabaseBackend$getSuggestedControlCondtions()}}
\item \href{#method-getSuggestedControlIngredients}{\code{CemDatabaseBackend$getSuggestedControlIngredients()}}
\item \href{#method-clone}{\code{CemDatabaseBackend$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
initialize backend object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemDatabaseBackend$new(
  connectionDetails,
  cemSchema,
  vocabularySchema,
  sourceSchema,
  usePooledConnection = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{connectionDetails}}{DatabaseConnector connection details object}

\item{\code{cemSchema}}{Schema name including CEM unified and matrix_summary tables}

\item{\code{vocabularySchema}}{OMOP vocabulary}

\item{\code{sourceSchema}}{Schema containing CEM source information table}

\item{\code{usePooledConnection}}{Used a pooled connection object rather than a database connector object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Closes connection
Condition evidence summary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemDatabaseBackend$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getConditionEvidenceSummary"></a>}}
\if{latex}{\out{\hypertarget{method-getConditionEvidenceSummary}{}}}
\subsection{Method \code{getConditionEvidenceSummary()}}{
Reutrns set of ingredient concepts for a given conceptset of outcomes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemDatabaseBackend$getConditionEvidenceSummary(
  conditionConceptSet,
  siblingLookupLevels = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{conditionConceptSet}}{data.frame conforming to conceptset format, must be standard SNOMED conditions}

\item{\code{siblingLookupLevels}}{where mapping is not found it may be beneficial to lookup siblings in the concept ancestry. This defines the number of levels to jump
Condition evidence}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getConditionEvidence"></a>}}
\if{latex}{\out{\hypertarget{method-getConditionEvidence}{}}}
\subsection{Method \code{getConditionEvidence()}}{
Reutrns set of relationships that exist within CEM for a condition conceptset of interest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemDatabaseBackend$getConditionEvidence(
  conditionConceptSet,
  siblingLookupLevels = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{conditionConceptSet}}{data.frame conforming to conceptset format, must be standard SNOMED conditions}

\item{\code{siblingLookupLevels}}{where mapping is not found it may be beneficial to lookup siblings in the concept ancestry. This defines the number of levels to jump
Ingredient evidence summary}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getIngredientEvidenceSummary"></a>}}
\if{latex}{\out{\hypertarget{method-getIngredientEvidenceSummary}{}}}
\subsection{Method \code{getIngredientEvidenceSummary()}}{
Reutrns set of outcome concepts for a given conceptset of ingredients/exposures
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemDatabaseBackend$getIngredientEvidenceSummary(ingredientConceptSet)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ingredientConceptSet}}{data.frame conforming to conceptset format, must be standard RxNorm Ingredients
Ingredient conceptset evidence}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getIngredientEvidence"></a>}}
\if{latex}{\out{\hypertarget{method-getIngredientEvidence}{}}}
\subsection{Method \code{getIngredientEvidence()}}{
for a concept set of rxnorm ingredients (or their ancestors) list the evidence stored in the CEM
Utilises matrix summary table for optimisation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemDatabaseBackend$getIngredientEvidence(ingredientConceptSet)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ingredientConceptSet}}{data.frame conforming to conceptset format, must be standard RxNorm Ingredients}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getRelationships"></a>}}
\if{latex}{\out{\hypertarget{method-getRelationships}{}}}
\subsection{Method \code{getRelationships()}}{
From a unified CEM relationships table, for a conceptSet of drug ingredients and a conceptSet of conditions,
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemDatabaseBackend$getRelationships(
  ingredientConceptSet,
  conditionConceptSet,
  conditionSiblingLookupLevels = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ingredientConceptSet}}{data.frame conforming to conceptset format, must be standard RxNorm Ingredients}

\item{\code{conditionConceptSet}}{data.frame conforming to conceptset format, must be standard SNOMED conditions}

\item{\code{conditionSiblingLookupLevels}}{integer - where mapping is not found it may be beneficial to lookup siblings in the concept ancestry. This defines the number of levels to jump}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCemSourceInfo"></a>}}
\if{latex}{\out{\hypertarget{method-getCemSourceInfo}{}}}
\subsection{Method \code{getCemSourceInfo()}}{
Get CEM source info as a dataframe
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemDatabaseBackend$getCemSourceInfo()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSuggestedControlCondtions"></a>}}
\if{latex}{\out{\hypertarget{method-getSuggestedControlCondtions}{}}}
\subsection{Method \code{getSuggestedControlCondtions()}}{
Get negative control snomed condition concepts for a given conceptset
These are ranked by co-occurence accross ohdsi studies
A negative control for a submitted concept_set is valid if there is no evidence for the outcome
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemDatabaseBackend$getSuggestedControlCondtions(
  ingredientConceptSet,
  nControls = 50
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ingredientConceptSet}}{data.frame conforming to conceptset format, must be standard RxNorm Ingredients}

\item{\code{nControls}}{topN controls to select - the maximum number will be limited by available concepts without related evidence}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSuggestedControlIngredients"></a>}}
\if{latex}{\out{\hypertarget{method-getSuggestedControlIngredients}{}}}
\subsection{Method \code{getSuggestedControlIngredients()}}{
Get negative control rxnorm ingredient concepts for a given conceptset
These are ranked by co-occurence accross ohdsi studies
A negative control for a submitted concept_set is valid if there is no evidence for the ingredient/condition combination
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemDatabaseBackend$getSuggestedControlIngredients(
  conditionConceptSet,
  siblingLookupLevels = 0,
  nControls = 50
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{conditionConceptSet}}{data.frame conforming to conceptset format, must be standard SNOMED conditions}

\item{\code{siblingLookupLevels}}{where mapping is not found it may be beneficial to lookup siblings in the concept ancestry. This defines the number of levels to jump}

\item{\code{nControls}}{topN controls to select - the maximum number will be limited by available concepts without related evidence}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemDatabaseBackend$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
