% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebBackend.R
\name{CemWebApiBackend}
\alias{CemWebApiBackend}
\title{CEM Web Backend Class}
\description{
An interface to the common evidence model that uses http requests
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiUrl}}{url for the common evidence model hosted api instance}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CemWebApiBackend$new()}}
\item \href{#method-request}{\code{CemWebApiBackend$request()}}
\item \href{#method-getStatus}{\code{CemWebApiBackend$getStatus()}}
\item \href{#method-getConditionEvidenceSummary}{\code{CemWebApiBackend$getConditionEvidenceSummary()}}
\item \href{#method-getIngredientEvidenceSummary}{\code{CemWebApiBackend$getIngredientEvidenceSummary()}}
\item \href{#method-getConditionEvidence}{\code{CemWebApiBackend$getConditionEvidence()}}
\item \href{#method-getIngredientEvidence}{\code{CemWebApiBackend$getIngredientEvidence()}}
\item \href{#method-getRelationships}{\code{CemWebApiBackend$getRelationships()}}
\item \href{#method-getCemSourceInfo}{\code{CemWebApiBackend$getCemSourceInfo()}}
\item \href{#method-getVersion}{\code{CemWebApiBackend$getVersion()}}
\item \href{#method-clone}{\code{CemWebApiBackend$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
initialzie object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemWebApiBackend$new(apiUrl)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{apiUrl}}{String URL parameter for hosted}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-request"></a>}}
\if{latex}{\out{\hypertarget{method-request}{}}}
\subsection{Method \code{request()}}{
Do a web request
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemWebApiBackend$request(method, endpoint, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{string "POST", "GET", "PUT" (not implemented), "DELETE" (not implemented)}

\item{\code{endpoint}}{URL endpoint string}

\item{\code{...}}{list params for httr method}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getStatus"></a>}}
\if{latex}{\out{\hypertarget{method-getStatus}{}}}
\subsection{Method \code{getStatus()}}{
GET server status
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemWebApiBackend$getStatus()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getConditionEvidenceSummary"></a>}}
\if{latex}{\out{\hypertarget{method-getConditionEvidenceSummary}{}}}
\subsection{Method \code{getConditionEvidenceSummary()}}{
Reutrns set of ingredient concepts for a given conceptset of outcomes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemWebApiBackend$getConditionEvidenceSummary(
  conditionConceptSet,
  siblingLookupLevels = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{conditionConceptSet}}{data.frame conforming to conceptset format, must be standard SNOMED conditions}

\item{\code{siblingLookupLevels}}{where mapping is not found it may be beneficial to lookup siblings in the concept ancestry. This defines the number of levels to jump}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getIngredientEvidenceSummary"></a>}}
\if{latex}{\out{\hypertarget{method-getIngredientEvidenceSummary}{}}}
\subsection{Method \code{getIngredientEvidenceSummary()}}{
Reutrns set of outcome concepts for a given conceptset of ingredients
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemWebApiBackend$getIngredientEvidenceSummary(ingredientConceptSet)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ingredientConceptSet}}{data.frame conforming to conceptset format, must be standard RxNorm Ingredients}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getConditionEvidence"></a>}}
\if{latex}{\out{\hypertarget{method-getConditionEvidence}{}}}
\subsection{Method \code{getConditionEvidence()}}{
Reutrns ingredient evidence for a given conceptset of conditions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemWebApiBackend$getConditionEvidence(
  conditionConceptSet,
  siblingLookupLevels = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{conditionConceptSet}}{data.frame conforming to conceptset format, must be standard SNOMED conditions}

\item{\code{siblingLookupLevels}}{where mapping is not found it may be beneficial to lookup siblings in the concept ancestry. This defines the number of levels to jump}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getIngredientEvidence"></a>}}
\if{latex}{\out{\hypertarget{method-getIngredientEvidence}{}}}
\subsection{Method \code{getIngredientEvidence()}}{
Reutrns condtion evidence for a given conceptset of ingredients
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemWebApiBackend$getIngredientEvidence(ingredientConceptSet)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ingredientConceptSet}}{data.frame conforming to conceptset format, must be standard RxNorm Ingredients}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getRelationships"></a>}}
\if{latex}{\out{\hypertarget{method-getRelationships}{}}}
\subsection{Method \code{getRelationships()}}{
From a unified CEM relationships table, for a conceptSet of drug ingredients and a conceptSet of conditions,
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemWebApiBackend$getRelationships(
  ingredientConceptSet,
  conditionConceptSet,
  conditionSiblingLookupLevels = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ingredientConceptSet}}{data.frame conforming to conceptset format, must be standard RxNorm Ingredients}

\item{\code{conditionConceptSet}}{data.frame conforming to conceptset format, must be standard SNOMED conditions}

\item{\code{conditionSiblingLookupLevels}}{integer - where mapping is not found it may be beneficial to lookup siblings in the concept ancestry. This defines the number of levels to jump}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCemSourceInfo"></a>}}
\if{latex}{\out{\hypertarget{method-getCemSourceInfo}{}}}
\subsection{Method \code{getCemSourceInfo()}}{
Returns datframe of sources that made the CEM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemWebApiBackend$getCemSourceInfo()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getVersion"></a>}}
\if{latex}{\out{\hypertarget{method-getVersion}{}}}
\subsection{Method \code{getVersion()}}{
getVersion information from api
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemWebApiBackend$getVersion()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemWebApiBackend$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
