% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{loadApi}
\alias{loadApi}
\title{Load API}
\usage{
loadApi(
  connectionDetails,
  cemSchema = Sys.getenv("CEM_DATABASE_SCHEMA"),
  vocabularySchema = Sys.getenv("CEM_DATABASE_VOCAB_SCHEMA"),
  sourceSchema = Sys.getenv("CEM_DATABASE_INFO_SCHEMA"),
  pathToPlumberApi = system.file(file.path("api", "plumber.R"), package =
    "CEMConnector"),
  envir = new.env(parent = .GlobalEnv)
)
}
\value{
Plumber router object
}
\description{
Loads plumber API
}
