% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvidenceExplorer.R
\name{negativeControlSelectorModule}
\alias{negativeControlSelectorModule}
\title{Negative control selection  utility}
\usage{
negativeControlSelectorModule(
  id,
  backend,
  conceptInput = NULL,
  siblingLookupLevelsInput = shiny::reactive({     0 }),
  isOutcomeSearch = shiny::reactive({     TRUE }),
  nControls = shiny::reactive({     50 })
)
}
\arguments{
\item{id}{string - unique namespace for module}

\item{backend}{CemConnector backend (database or Api URL)}

\item{conceptInput}{shiny::reactive that returns data.frame with headers conceptId, includeDescendants, isExcluded}

\item{siblingLookupLevelsInput}{shiny::reactive that returns positive integer for sibling levels to lookup for condition concept mappings to CEM}

\item{nControls}{shiny::reactive that returns positive integer for number of controls to get}

\item{defaultToOutcomeControls}{boolean}
}
\description{
Shiny Module for integration of evidence for selecting negative controls conceptsets in to shiny applications
}
