% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CemConnector.R
\docType{package}
\name{CemConnector-package}
\alias{CemConnector-package}
\title{CemConnector: OHDSI Common Evidence Model (CEM) Connector}
\description{
RESTFul API and Database Backend connections for the OHDSI Common Evidence Model Built around an R Plumber api.
}
\keyword{internal}
