% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebBackend.R
\name{CemWebApiBackend}
\alias{CemWebApiBackend}
\title{CEM Web Backend Class}
\value{
data.frame of condition concept_id and concept_name

data.frame of condition concept_id and concept_name
}
\description{
An interface to the common evidence model that uses http requests
}
\section{Super class}{
\code{\link[CemConnector:AbstractCemBackend]{CemConnector::AbstractCemBackend}} -> \code{CemWebApiBackend}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiUrl}}{url for the common evidence model hosted api instance}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CemWebApiBackend$new()}}
\item \href{#method-request}{\code{CemWebApiBackend$request()}}
\item \href{#method-getStatus}{\code{CemWebApiBackend$getStatus()}}
\item \href{#method-getConditionEvidenceSummary}{\code{CemWebApiBackend$getConditionEvidenceSummary()}}
\item \href{#method-getIngredientEvidenceSummary}{\code{CemWebApiBackend$getIngredientEvidenceSummary()}}
\item \href{#method-getConditionEvidence}{\code{CemWebApiBackend$getConditionEvidence()}}
\item \href{#method-getIngredientEvidence}{\code{CemWebApiBackend$getIngredientEvidence()}}
\item \href{#method-getRelationships}{\code{CemWebApiBackend$getRelationships()}}
\item \href{#method-getCemSourceInfo}{\code{CemWebApiBackend$getCemSourceInfo()}}
\item \href{#method-getVersion}{\code{CemWebApiBackend$getVersion()}}
\item \href{#method-getSuggestedControlCondtions}{\code{CemWebApiBackend$getSuggestedControlCondtions()}}
\item \href{#method-getSuggestedControlIngredients}{\code{CemWebApiBackend$getSuggestedControlIngredients()}}
\item \href{#method-clone}{\code{CemWebApiBackend$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="CemConnector" data-topic="AbstractCemBackend" data-id="finalize">}\href{../../CemConnector/html/AbstractCemBackend.html#method-finalize}{\code{CemConnector::AbstractCemBackend$finalize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
initialize object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemWebApiBackend$new(apiUrl)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{apiUrl}}{String URL parameter for hosted}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-request"></a>}}
\if{latex}{\out{\hypertarget{method-request}{}}}
\subsection{Method \code{request()}}{
Do a web request
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemWebApiBackend$request(method, endpoint, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{string "POST", "GET", "PUT" (not implemented), "DELETE" (not implemented)}

\item{\code{endpoint}}{URL endpoint string}

\item{\code{...}}{list params for httr method}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getStatus"></a>}}
\if{latex}{\out{\hypertarget{method-getStatus}{}}}
\subsection{Method \code{getStatus()}}{
GET server status
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemWebApiBackend$getStatus()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getConditionEvidenceSummary"></a>}}
\if{latex}{\out{\hypertarget{method-getConditionEvidenceSummary}{}}}
\subsection{Method \code{getConditionEvidenceSummary()}}{
Returns set of ingredient concepts for a given conceptset of outcomes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemWebApiBackend$getConditionEvidenceSummary(
  conditionConceptSet,
  siblingLookupLevels = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{conditionConceptSet}}{data.frame conforming to conceptset format, must be standard SNOMED conditions}

\item{\code{siblingLookupLevels}}{where mapping is not found it may be beneficial to lookup siblings in the concept ancestry. This defines the number of levels to jump}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getIngredientEvidenceSummary"></a>}}
\if{latex}{\out{\hypertarget{method-getIngredientEvidenceSummary}{}}}
\subsection{Method \code{getIngredientEvidenceSummary()}}{
Returns set of outcome concepts for a given conceptset of ingredients
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemWebApiBackend$getIngredientEvidenceSummary(ingredientConceptSet)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ingredientConceptSet}}{data.frame conforming to conceptset format, must be standard RxNorm Ingredients}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getConditionEvidence"></a>}}
\if{latex}{\out{\hypertarget{method-getConditionEvidence}{}}}
\subsection{Method \code{getConditionEvidence()}}{
Returns ingredient evidence for a given conceptset of conditions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemWebApiBackend$getConditionEvidence(
  conditionConceptSet,
  siblingLookupLevels = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{conditionConceptSet}}{data.frame conforming to conceptset format, must be standard SNOMED conditions}

\item{\code{siblingLookupLevels}}{where mapping is not found it may be beneficial to lookup siblings in the concept ancestry. This defines the number of levels to jump}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getIngredientEvidence"></a>}}
\if{latex}{\out{\hypertarget{method-getIngredientEvidence}{}}}
\subsection{Method \code{getIngredientEvidence()}}{
Returns condtion evidence for a given conceptset of ingredients
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemWebApiBackend$getIngredientEvidence(ingredientConceptSet)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ingredientConceptSet}}{data.frame conforming to conceptset format, must be standard RxNorm Ingredients}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getRelationships"></a>}}
\if{latex}{\out{\hypertarget{method-getRelationships}{}}}
\subsection{Method \code{getRelationships()}}{
From a unified CEM relationships table, for a conceptSet of drug ingredients and a conceptSet of conditions,
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemWebApiBackend$getRelationships(
  ingredientConceptSet,
  conditionConceptSet,
  conditionSiblingLookupLevels = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ingredientConceptSet}}{data.frame conforming to conceptset format, must be standard RxNorm Ingredients}

\item{\code{conditionConceptSet}}{data.frame conforming to conceptset format, must be standard SNOMED conditions}

\item{\code{conditionSiblingLookupLevels}}{integer - where mapping is not found it may be beneficial to lookup siblings in the concept ancestry. This defines the number of levels to jump}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCemSourceInfo"></a>}}
\if{latex}{\out{\hypertarget{method-getCemSourceInfo}{}}}
\subsection{Method \code{getCemSourceInfo()}}{
Returns data.frame of sources that made the CEM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemWebApiBackend$getCemSourceInfo()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getVersion"></a>}}
\if{latex}{\out{\hypertarget{method-getVersion}{}}}
\subsection{Method \code{getVersion()}}{
getVersion information from api
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemWebApiBackend$getVersion()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSuggestedControlCondtions"></a>}}
\if{latex}{\out{\hypertarget{method-getSuggestedControlCondtions}{}}}
\subsection{Method \code{getSuggestedControlCondtions()}}{
Get negative control snomed condition concepts for a given conceptset
These are ranked by co-occurrence accross ohdsi studies
A negative control for a submitted concept_set is valid if there is no evidence for the outcome
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemWebApiBackend$getSuggestedControlCondtions(
  ingredientConceptSet,
  nControls = 50
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ingredientConceptSet}}{data.frame conforming to conceptset format, must be standard RxNorm Ingredients}

\item{\code{nControls}}{topN controls to select - the maximum number will be limited by available concepts without related evidence}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSuggestedControlIngredients"></a>}}
\if{latex}{\out{\hypertarget{method-getSuggestedControlIngredients}{}}}
\subsection{Method \code{getSuggestedControlIngredients()}}{
Get negative control rxnorm ingredient concepts for a given conceptset
These are ranked by co-occurrence accross ohdsi studies
A negative control for a submitted concept_set is valid if there is no evidence for the ingredient/condition combination
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemWebApiBackend$getSuggestedControlIngredients(
  conditionConceptSet,
  siblingLookupLevels = 0,
  nControls = 50
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{conditionConceptSet}}{data.frame conforming to conceptset format, must be standard SNOMED conditions}

\item{\code{siblingLookupLevels}}{where mapping is not found it may be beneficial to lookup siblings in the concept ancestry. This defines the number of levels to jump}

\item{\code{nControls}}{topN controls to select - the maximum number will be limited by available concepts without related evidence}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CemWebApiBackend$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
