% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{loadApi}
\alias{loadApi}
\title{Load API}
\usage{
loadApi(
  connectionDetails,
  cemSchema = Sys.getenv("CEM_DATABASE_SCHEMA"),
  vocabularySchema = Sys.getenv("CEM_DATABASE_VOCAB_SCHEMA"),
  sourceSchema = Sys.getenv("CEM_DATABASE_INFO_SCHEMA"),
  pathToPlumberApi = system.file(file.path("api", "plumber.R"), package =
    "CemConnector"),
  envir = new.env(parent = .GlobalEnv)
)
}
\arguments{
\item{connectionDetails}{DatabaseConnector connection details object}

\item{cemSchema}{schema where matrix summary and merged evidence are found}

\item{vocabularySchema}{vocabulary schema on database}

\item{sourceSchema}{schema for info about the CEM}

\item{pathToPlumberApi}{path to plumber script (default is package's)}

\item{envir}{R environment}
}
\value{
Plumber router object
}
\description{
Loads plumber API for functions
}
