% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvidenceExplorer.R
\name{negativeControlSelectorUi}
\alias{negativeControlSelectorUi}
\title{negative control module ui}
\usage{
negativeControlSelectorUi(id)
}
\arguments{
\item{id}{string - unique namespace for module. Must match call to ceExplorerModule}
}
\description{
Shiny Module for integration of negative control search for conceptsets in to shiny applications
}
